% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_scrape_markdown.R
\name{scrape_markdown}
\alias{scrape_markdown}
\title{A Scraping Markdown Function}
\usage{
scrape_markdown(markdown)
}
\arguments{
\item{markdown}{character string of markdown format, containing flashcard info}
}
\description{
This function converts passed characters from a markdown file, into a table of flashcards
}
\examples{
scrape_markdown()
}
\keyword{anki}
\keyword{markdown}
