% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_scrape_markdown_file.R
\name{scrape_markdown_file}
\alias{scrape_markdown_file}
\title{A Scraping Markdown File Function}
\usage{
scrape_markdown_file(markdown_file)
}
\arguments{
\item{markdown_file}{path pointing to a markdown file}
}
\description{
This function converts a markdown file, into a table of flashcards
}
\examples{
scrape_markdown_file()
}
\keyword{anki}
\keyword{markdown}
